Corrected the author name and fixed the reducer,driver code.

+++ src/main/java/com/techidiocy/hadoop/mapreduce/drivers/FlipkartCorrelationCoeffDriver.java
@@ -5,7 +5,7 @@ package com.techidiocy.hadoop.mapreduce.drivers;
 
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -39,7 +39,7 @@ public class FlipkartCorrelationCoeffDriver extends Configured implements Tool {
 		job.setMapOutputValueClass(Item.class);
 		
 		job.setOutputKeyClass(Text.class);
-		job.setOutputValueClass(LongWritable.class);
+		job.setOutputValueClass(DoubleWritable.class);
 		
 		job.setInputFormatClass(KeyValueTextInputFormat.class);
